feat: добавлены компоненты InfoBlock и PhotoBlock, добавлен memo в экспорте
