+++ src/components/ProjectPopup/PhotoBlock/styles.module.scss
@@ -1,0 +1,66 @@
+@import 'styles/global';
+
+.photoBlock {
+  @include flex(column, center, center);
+
+  position:relative;
+
+  width: 50%;
+  height: 100%;
+
+  opacity: 1;
+  background: $color-another-main-text;
+  border-radius: 0 2px 2px 0;
+}
+
+.hidden {
+  display: none;
+}
+
+.carouselItem {
+  width: 100%;
+  max-height: 100%;
+  padding: 8px;
+
+  border-radius: 2px;
+}
+
+.button {
+  cursor: pointer;
+
+  width: 56px;
+  height: 56px;
+  padding: 0;
+
+  font: inherit;
+  color: inherit;
+
+  background: none;
+  border: none;
+  outline: inherit;
+}
+
+.buttons {
+  @include flex(row, space-between, center);
+
+  position: absolute;
+  z-index: 2;
+  top: 50%;
+  right: 0;
+  left: 0;
+  transform: translateY(-50%);
+
+  width: 100%;
+  padding: 0 2px;
+}
+
+.alice-carousel {
+  z-index: 1;
+}
+
+@media (max-width: 1024px) {
+  .photoBlock {
+    width:100%;
+    height: 50%;
+  }
+}
