+++ src/types/IPhotoBlockProps.ts
@@ -1,0 +1,3 @@
+export interface IPhotoBlockProps {
+  imagesUrls: string[] | null;
+}
